Fix: Back to working state after some variations

+++ src/main/java/com/example/demo/Models/User.java
@@ -35,7 +35,7 @@ public class User{
 
     }
 
-    public Long getId() {
+    public Integer getId() {
         return id;
     }
     public void setId(Integer id) {
